Add a helpful hint for users of Data.Compact
Compaction traverses the whole of a data structure, including in the case of a Vector, the extra bottoms that might be hidden in the unused elements at the end of an oversized array that was allocated by `fromList`. To help people who trip over this gotcha, let's add a hint to the error message.